+++ src/Prokudin.Gui/Services/DistributionInfo.cs
@@ -1,0 +1,7 @@
+namespace Prokudin.Gui.Services;
+
+public static class DistributionInfo
+{
+    public const string GitHubOwner = "antsht";
+    public const string GitHubRepo = "-Prokudin";
+}